// src/app/features/star-admin/models/star.ts
export interface Star {
  id: string
  edr3Id: number
  name: string
  ra: string
  pmRa: number
  dec: string
  pmDec: number
  parallax: number
  luminosity: number
  radius: number
  mass: number
  temperature: number
  magnitude: number
  spectralType: string
  evolutionaryStage: string
  age: number
}